enable -Werror for host tools and fix all warnings
  Add -Werror to HOST_CFLAGS to catch warnings at compile time.
  Fix all existing warnings in cmd/cc/:

  - Sign-compare: use size_t for array iteration, add casts where needed
  - Missing field initializers: use designated initializers for Type and
  InitDesg compound literals
  - Unhandled switch cases: add default cases to eval2, eval_rval,
  is_const_expr, eval_double, is_compatible, and add_type
  - Integer comparison: cast -1 to uint32_t in unicode range check

+++ cmd/cc/hashmap.c
@@ -159,19 +159,19 @@ void hashmap_test(void) {
 	}
 
 	for (int i = 0; i < 1000; i++) {
-		assert((size_t)hashmap_get(map, format("key %d", i)) == i);
+		assert((size_t)hashmap_get(map, format("key %d", i)) == (size_t)i);
 	}
 	for (int i = 1000; i < 1500; i++) {
 		assert(hashmap_get(map, "no such key") == NULL);
 	}
 	for (int i = 1500; i < 1600; i++) {
-		assert((size_t)hashmap_get(map, format("key %d", i)) == i);
+		assert((size_t)hashmap_get(map, format("key %d", i)) == (size_t)i);
 	}
 	for (int i = 1600; i < 2000; i++) {
 		assert(hashmap_get(map, "no such key") == NULL);
 	}
 	for (int i = 2000; i < 5000; i++) {
-		assert((size_t)hashmap_get(map, format("key %d", i)) == i);
+		assert((size_t)hashmap_get(map, format("key %d", i)) == (size_t)i);
 	}
 	for (int i = 5000; i < 6000; i++) {
 		assert(hashmap_get(map, "no such key") == NULL);
